修复后台管理当对应管理无东西时，查看不了对应页面的bug

+++ blog/internal/middleware/auth.go
@@ -100,8 +100,8 @@ func GetUsername(c *gin.Context) string {
 // GetRoleID 从上下文获取角色ID
 func GetRoleID(c *gin.Context) int64 {
 	if roleID, exists := c.Get(ContextRoleID); exists {
-		if id, ok := roleID.(int64); ok {
-			return id
+		if id, ok := roleID.(uint); ok {
+			return int64(id)
 		}
 	}
 	return 0
